Restructured the overall class structure to allow easier setup of motors using.
A new function is provided in logger (purely as a convenient place to put it) which
finds a class and calls its constructor with all other params.

All the other class' updated to work this way.

The 3 example config files for dc_h_bridge configurations have been removed and
replaced with a single file config_h_bridge.py.