2 big areas of change here.
First I have ditched the character mode app that runs the motor through ssh / terminal session, and replaced
it with a web browser based front end (that lives in a separate library). This is why 3 files are deleted.

Secondly, after various discussions on the Raspberry pi forums I have significantly re-designed the way
quad encoder feedback is handled. See the discussion at
https://www.raspberrypi.org/forums/viewtopic.php?f=37&t=213610
There is now a small C program (quadfeedback.c) that runs in a separate process and uses the C interface
to the pigpio daemon. This uses shared memory (via an mmap'ed file) to allow quadfast.py to poll on demand
for the latest known position (quadfast.py runs at motorset level for efficiency). This in turn is accessed
by quadfastencoder.py which can be used in place of quadencoder.py.

A few other small changes to support the web based front end are included.